Fix the automatic unsubscription of CallOnce handlers
Before this commit the automatic unsubscription of a CallOnce handler
was subject to a race between the goroutine which closes a handler
when it is unsubscribed and the writing of the event in the queue of
the handler. If the race was won by the goroutine closing the handler,
the event was dropped. Now the event is put in the handler queue
before the automatic unsubscription of the handler.